pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "JEPA-WMs"
version = "1.0.0"
readme = "README.md"
requires-python = ">=3.10,<3.11"
authors = [
    {name = "Fundamental AI Research"}
]
keywords = ["machine learning", "world models", "jepa", "reinforcement learning"]

dependencies = [
    # Core dependencies
    "torch>=2.7.0",
    "torchvision==0.22.0",
    "torchrl>=0.9.2",
    "tensordict>=0.9.1",

    # Vision and transformers
    "timm>=1.0.19",

    # Data processing
    "numpy",
    # "numpy==1.23.3",
    "opencv-python",
    "pillow",
    "decord>=0.6.0",
    "h5py",
    "pandas",
    "einops",

    # Scientific computing
    "scipy",
    "scikit-image",

    # Visualization
    "matplotlib",
    "seaborn",
    "plotly",
    "mediapy",
    "imageio",
    "imageio-ffmpeg",
    "moviepy",
    "termcolor",

    # Configuration management
    "omegaconf",

    # ML experiment tracking and logging
    "wandb",

    # Utilities
    "tqdm",
    "submitit",
    "clusterscope",

    # Robotics and simulation environments
    "gym==0.23.1",
    "gymnasium>=1.2.0",
    "pygame>=2.6.0",
    "pymunk==6.8.0",
    "shapely",
    "pybullet",
    "mujoco>=3.2.0",
    "dm-control>=1.0.34",

    # Metaworld (from git)
    "metaworld @ git+https://github.com/Farama-Foundation/Metaworld.git@master",

    # Additional utilities
    "ruamel.yaml",
]

[project.optional-dependencies]
# Development dependencies
dev = [
    "pytest>=8.4.0",
    "black>=25.0.0",
    "isort>=7.0.0",
    "flake8>=7.0.0",
    "ipykernel",
    "jupyterlab",
]

# Robot manipulation environments
# NOTE: robocasa and robosuite must be installed manually from source.
# See the "Optional: Robocasa" section in README.md for installation instructions.
# These packages cannot be installed via pip alone as they require:
# - Cloning from specific git repositories:
#   - https://github.com/Basile-Terv/robosuite (robocasa-dev branch)
#   - https://github.com/Basile-Terv/robocasa
# - Installing with pip install -e .
# - Downloading additional assets (~5GB) for robocasa
robocasa = [
    # "robocasa",  # Install manually - see README.md
    # "robosuite",  # Install manually - see README.md
]

# Full environment with all extras
all = [
    "jepa[dev,robocasa]",
]

[project.urls]
Homepage = "https://github.com/fairinternal/jepa-internal"
Repository = "https://github.com/fairinternal/jepa-internal"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "app*", "evals*"]

[tool.isort]
profile = "black"
line_length = 119

[tool.black]
line-length = 119

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
